fix migration issues when running db:initial_setup
closes CNVS-38243

test plan:
* In your environment, run the following rake commands:

  bundle exec rake db:drop db:create
  bundle exec rake db:initial_setup

* Verify that the db:initial_setup command successfully completes
  all migrations and begins to ask your site administrator
  account details

Reviewed-on: https://gerrit.instructure.com/119244
Tested-by: Jenkins
QA-Review: Matt Taylor <mtaylor@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>

+++ db/migrate/20170707212935_drop_old_collation_key_indexes.rb
@@ -24,8 +24,13 @@ class DropOldCollationKeyIndexes < ActiveRecord::Migration[5.0]
   def up
     return unless connection.extension_installed?(:pg_collkey)
 
-    remove_index :users, name: :index_users_on_sortable_name_old
-    remove_index :attachments, name: :index_attachments_on_folder_id_and_file_state_and_display_name1
+    if connection.index_name_exists?(:users, :index_users_on_sortable_name_old)
+      remove_index :users, name: :index_users_on_sortable_name_old
+    end
+
+    if connection.index_name_exists?(:attachments, :index_attachments_on_folder_id_and_file_state_and_display_name1)
+      remove_index :attachments, name: :index_attachments_on_folder_id_and_file_state_and_display_name1
+    end
   end
 
   def down
